[NEW] Separate build scripts as a module.
When a developer develops their own ROM, they need to depends on
zamgba. Unlike regular cases, they need to depends on build scripts
to provide a correct build target. We moved the code out from build.zig,
make it a module, so we can expose it later to user side.